# public/assets/api.yaml
openapi: 3.0.0
info:
  title: 'My First API'
  version: '0.1'
paths:
...
  /permissions:
    get:
      tags:
        - Permissions
      summary: 'Find list Permissions'
      description: 'Returns list of Permissions'
      operationId: getPermissions
      parameters:
        -
          name: search
          in: query
          description: 'search by column defined'
          schema:
            type: object
        -
          name: order
          in: query
          description: 'order by column defined'
          schema:
            type: object
        -
          name: page
          in: query
          description: 'page to show'
          schema:
            type: int32
        -
          name: limit
          in: query
          description: 'count data display per page'
          schema:
            type: int32
      responses:
        '200':
          description: 'successful operation'
          content:
            application/json:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Permissions' } }
                  pagination: { properties: { currentPage: { type: integer }, totalPage: { type: integer } }, type: object }
                type: object
            application/xml:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Permissions' } }
                  pagination: { type: array, items: { $ref: '#/components/schemas/Permissions' } }
                type: object
        '404':
          description: 'Permissions not found'
      security:
        -
          bearer_auth: []
    post:
      tags:
        - Permissions
      summary: 'Add a new Permissions to the store'
      operationId: addPermissions
      requestBody:
        $ref: '#/components/requestBodies/Permissions'
      responses:
        '201':
          description: 'Created Permissions'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Permissions'
            application/xml:
              schema:
                $ref: '#/components/schemas/Permissions'
        '405':
          description: 'Invalid input'
      security:
        -
          bearer_auth: []
  '/permissions/{id}':
    get:
      tags:
        - Permissions
      summary: 'Find Permissions by ID'
      description: 'Returns a single Permissions'
      operationId: getPermissionsById
      parameters:
        -
          name: id
          in: path
          description: 'ID of Permissions to return'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Permissions'
            application/xml:
              schema:
                $ref: '#/components/schemas/Permissions'
        '400':
          description: 'Invalid ID supplier'
        '404':
          description: 'Permissions not found'
      security:
        -
          bearer_auth: []
    put:
      tags:
        - Permissions
      summary: 'Update an existing Permissions'
      operationId: updatePermissions
      parameters:
        -
          name: id
          in: path
          description: 'Permissions id to update'
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        $ref: '#/components/requestBodies/Permissions'
      responses:
        '400':
          description: 'Invalid ID supplied'
        '404':
          description: 'Permissions not found'
        '405':
          description: 'Validation exception'
      security:
        -
          bearer_auth: []
    delete:
      tags:
        - Permissions
      summary: 'Deletes a Permissions'
      operationId: deletePermissions
      parameters:
        -
          name: id
          in: path
          description: 'Permissions id to delete'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '400':
          description: 'Invalid ID supplied'
        '404':
          description: 'Pet not found'
      security:
        -
          bearer_auth: []
  /posts:
    get:
      tags:
        - Posts
      summary: 'Find list Posts'
      description: 'Returns list of Posts'
      operationId: getPosts
      parameters:
        -
          name: search
          in: query
          description: 'search by column defined'
          schema:
            type: object
        -
          name: order
          in: query
          description: 'order by column defined'
          schema:
            type: object
        -
          name: page
          in: query
          description: 'page to show'
          schema:
            type: int32
        -
          name: limit
          in: query
          description: 'count data display per page'
          schema:
            type: int32
      responses:
        '200':
          description: 'successful operation'
          content:
            application/json:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Posts' } }
                  pagination: { properties: { currentPage: { type: integer }, totalPage: { type: integer } }, type: object }
                type: object
            application/xml:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Posts' } }
                  pagination: { type: array, items: { $ref: '#/components/schemas/Posts' } }
                type: object
        '404':
          description: 'Posts not found'
      security:
        -
          bearer_auth: []
    post:
      tags:
        - Posts
      summary: 'Add a new Posts to the store'
      operationId: addPosts
      requestBody:
        $ref: '#/components/requestBodies/Posts'
      responses:
        '201':
          description: 'Created Posts'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Posts'
            application/xml:
              schema:
                $ref: '#/components/schemas/Posts'
        '405':
          description: 'Invalid input'
      security:
        -
          bearer_auth: []
  '/posts/{id}':
    get:
      tags:
        - Posts
      summary: 'Find Posts by ID'
      description: 'Returns a single Posts'
      operationId: getPostsById
      parameters:
        -
          name: id
          in: path
          description: 'ID of Posts to return'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Posts'
            application/xml:
              schema:
                $ref: '#/components/schemas/Posts'
        '400':
          description: 'Invalid ID supplier'
        '404':
          description: 'Posts not found'
      security:
        -
          bearer_auth: []
    put:
      tags:
        - Posts
      summary: 'Update an existing Posts'
      operationId: updatePosts
      parameters:
        -
          name: id
          in: path
          description: 'Posts id to update'
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        $ref: '#/components/requestBodies/Posts'
      responses:
        '400':
          description: 'Invalid ID supplied'
        '404':
          description: 'Posts not found'
        '405':
          description: 'Validation exception'
      security:
        -
          bearer_auth: []
    delete:
      tags:
        - Posts
      summary: 'Deletes a Posts'
      operationId: deletePosts
      parameters:
        -
          name: id
          in: path
          description: 'Posts id to delete'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '400':
          description: 'Invalid ID supplied'
        '404':
          description: 'Pet not found'
      security:
        -
          bearer_auth: []
...
components:
  schemas:
...
    Permissions:
      title: Permissions
      description: Permissions
      properties:
        id:
          title: id
          description: id
          type: integer
          format: '-'
          maxLength: 10
          nullable: false
        name:
          title: name
          description: name
          type: string
          format: '-'
          maxLength: 191
          nullable: false
        slug:
          title: slug
          description: slug
          type: string
          format: '-'
          maxLength: 191
          nullable: false
        description:
          title: description
          description: description
          type: string
          format: '-'
          maxLength: 191
          nullable: true
        model:
          title: model
          description: model
          type: string
          format: '-'
          maxLength: 191
          nullable: true
        created_at:
          title: created_at
          description: created_at
          type: string
          format: '-'
          nullable: true
        updated_at:
          title: updated_at
          description: updated_at
          type: string
          format: '-'
          nullable: true
        deleted_at:
          title: deleted_at
          description: deleted_at
          type: string
          format: '-'
          nullable: true
      type: object
    Posts:
      title: Posts
      description: Posts
      properties:
        id:
          title: id
          description: id
          type: integer
          format: '-'
          maxLength: 20
          nullable: false
        user_id:
          title: user_id
          description: user_id
          type: integer
          format: '-'
          maxLength: 20
          nullable: false
        title:
          title: title
          description: title
          type: string
          format: '-'
          maxLength: 255
          nullable: false
        body:
          title: body
          description: body
          type: string
          format: '-'
          nullable: false
        cover_image:
          title: cover_image
          description: cover_image
          type: string
          format: '-'
          maxLength: 255
          nullable: false
        created_at:
          title: created_at
          description: created_at
          type: string
          format: '-'
          nullable: true
        updated_at:
          title: updated_at
          description: updated_at
          type: string
          format: '-'
          nullable: true
      type: object
...
  requestBodies:
...
    Permissions:
      description: 'Permissions object that needs to be added'
      content:
        application/x-www-form-urlencoded:
          schema:
            $ref: '#/components/schemas/Permissions'
        application/xml:
          schema:
            $ref: '#/components/schemas/Permissions'
        application/json:
          schema:
            $ref: '#/components/schemas/Permissions'
    Posts:
      description: 'Posts object that needs to be added'
      content:
        application/x-www-form-urlencoded:
          schema:
            $ref: '#/components/schemas/Posts'
        application/xml:
          schema:
            $ref: '#/components/schemas/Posts'
        application/json:
          schema:
            $ref: '#/components/schemas/Posts'
...